Improved frequency normalization algorithm
The old frequency normalization algorithm did not produce good results.
This seems to have been because it was likely to round frequencies in an
unfavorable way.  Change it to an improved algorithm which produces
better results; that is, a symbol will be, on average, assigned a number
of states which better reflects its mathematical entropy.

As an added bonus, the new algorithm is faster because it requires
neither sorting the symbols nor performing floating point arithmetic.

Note that changing the way frequencies are normalized does not change the
on-disk format.

Results for compressing the Silesia corpus (211,941,764 bytes):

    Old algorithm: Compressed to 67,809,153 bytes in 4.820 seconds
    New algorithm: Compressed to 67,628,688 bytes in 4.800 seconds