Do not waste time searching for length 3 matches
Since the hash function always hashes 4 bytes, it was extremely rare, if
not impossible for the specific hash function and hash table size, for
the compressor to ever find a length 3 match.  So don't waste time
searching for such matches.  This improves the compressor performance
slightly, with no change in compression ratio.