Fix fse_normalize_freq() when passed alphabet with no used symbols
I discovered there was a small mistake in my new frequency normalization
code.  LZFSE can call the function with an alphabet from which no symbols
were used.  In that case, it does not matter which normalized frequencies
we choose, but we must at least avoid dividing by 0.

An example input I found that demonstrated this problem was 100,000,000
bytes of zeroes.